Got the app.config copying working

+++ NugetCommandTest/NugetPackage/tools/install.ps1
@@ -7,5 +7,7 @@ $projectPath = Split-Path -Parent $project.FullName
 $toolAppConfig = Join-Path $contentPath "App.config"
 $projAppConfig = Join-Path $projectPath "App.config"
 
-Copy-Item $toolAppConfig $projAppConfig
-$project.ProjectItems.AddFromFile($projAppConfig)
+if (-not (Test-Path $projAppConfig)) {
+	Copy-Item $toolAppConfig $projAppConfig
+	$project.ProjectItems.AddFromFile($projAppConfig)
+}
